rename: 타입 파일 명 수정
